refactor actions
some renaming and other minor refactoring
