--- app/_actions/index.ts
@@ -1,4 +1,0 @@
-export * from './addPost';
-export * from './reportPost';
-export * from './setSiteCookies';
-export * from './vote';
