+++ app/_components/voting.tsx
@@ -1,6 +1,6 @@
 import { ChevronDownIcon, ChevronUpIcon } from '@heroicons/react/24/outline';
 
-import { vote as voteAction } from '@/app/_actions';
+import { votePost } from '@/app/_actions/post/vote';
 
 type VotingProps = {
   postId: string;
@@ -15,7 +15,7 @@ export function Voting({ postId, upvotes, vote }: VotingProps) {
       <form>
         <input type="hidden" name="postId" value={postId} />
         <input type="hidden" name="upvote" value="true" />
-        <button className="flex flex-col items-center gap-1" formAction={voteAction}>
+        <button className="flex flex-col items-center gap-1" formAction={votePost}>
           <ChevronUpIcon className={`w-5 ${vote === true ? 'text-zinc-600' : null}`} />
         </button>
       </form>
@@ -23,7 +23,7 @@ export function Voting({ postId, upvotes, vote }: VotingProps) {
       <form>
         <input type="hidden" name="postId" value={postId} />
         <input type="hidden" name="upvote" value="false" />
-        <button className="flex flex-col items-center gap-1" formAction={voteAction}>
+        <button className="flex flex-col items-center gap-1" formAction={votePost}>
           <ChevronDownIcon className={`w-5 ${vote === false ? 'text-zinc-600' : null}`} />
         </button>
       </form>
